Types

YaXLSX.XLSX — Type

XLSX

Structure with the contents of the XLSX file returned by the parse_xlsx function.

Fields

workbook::Workbook: XLSX file content.

sheets::Vector{Sheet}: vector of book sheets.

shared_strings::Union{Nothing,sharedStrings}: shared strings content.

Accessors

xl_sheets(xl_book::XLSX) -> xl_book.sheets

YaXLSX.Sheet — Type

Sheet

Structure with the contents of the book sheets contained in XLSX structure.

Fields

sheetData::Rows: sheet data.

table::Union{Nothing, Matrix}: sheet data in matrix form.

name::Union{Nothing, String}: sheet name.

YaXLSX.RowTable — Type

RowTable <: AbstractVector{<:NamedTuple}

Tabular data type that accepts NamedTuple vector and implements a row table.

YaXLSX.ColumnTable — Type

ColumnTable <: NamedTuple

Tabular data type that accepts NamedTuple and implements a column table.

Parsing

YaXLSX.parse_xlsx — Function

parse_xlsx(x::AbstractString) -> XLSX
parse_xlsx(x::Vector{UInt8}) -> XLSX

Parse a XLSX file into a structure of type XLSX.

Examples

julia> xl_book = parse_xlsx(xlsx_simple_table())
XLSX with 1 sheet

Book interfaces

YaXLSX.xl_sheetnames — Function

xl_sheetnames(xl_book::XLSX)

Get a list of all Sheet names in a book.

julia> xl_book = parse_xlsx(xlsx_two_sheets_table())
xXLSX with 2 sheets

julia> xl_sheetnames(xl_book)
2-element Vector{String}:
 "Sheet1"
 "Sheet2"

YaXLSX.xl_sheets — Function

xl_sheets(xl_book::XLSX) -> Vector{Sheet}
xl_sheets(xl_book::XLSX, keys_set::Vector{String}) -> Vector{Sheet}
xl_sheets(xl_book::XLSX, keys_set::UnitRange{Int64}) -> Vector{Sheet}
xl_sheets(xl_book::XLSX, key::String) -> Sheet
xl_sheets(xl_book::XLSX, key::Int64) -> Sheet

Get a Sheet or Vector{Sheet} by key or set of keys.

julia> xl_book = parse_xlsx(xlsx_two_sheets_table())
xXLSX with 2 sheets

julia> xl_sheets(xl_book)
2-element Vector{Sheet}:
 Sheet("Sheet1")
 Sheet("Sheet2")

julia> xl_sheets(xl_book, ["Sheet1", "Sheet2"])
2-element Vector{Sheet}:
 Sheet("Sheet1")
 Sheet("Sheet2")

julia> xl_sheets(xl_book, 1:2)
2-element Vector{Sheet}:
 Sheet("Sheet1")
 Sheet("Sheet2")

julia> xl_sheets(xl_book, "Sheet1")
Sheet("Sheet1")

julia> xl_sheets(xl_book, 1)
Sheet("Sheet1")

Sheet interfaces

YaXLSX.xl_rowtable — Function

xl_rowtable(sheet::Sheet, s::Union{UnitRange, AbstractString}; column_labels=nothing) -> RowTable
xl_rowtable(sheet::Sheet; column_labels=nothing) -> RowTable

Get a RowTable from Sheet by cells or columns range.

xl_book = parse_xlsx(xlsx_two_sheets_table())

xl_sheet = xl_sheets(xl_book, 2)

julia> xl_rowtable(xl_sheet, "A1:C3")
3-element Vector{NamedTuple}:
 (A = "Numbers", B = "Names", C = nothing)
 (A = 1.0, B = "a", C = nothing)
 (A = 2.0, B = "b", C = nothing)

julia> xl_rowtable(xl_sheet, "B:C"; column_labels=["c1", "c2"])
8-element Vector{NamedTuple}:
 (c1 = "Names", c2 = nothing)
 (c1 = "a", c2 = nothing)
 (c1 = "b", c2 = nothing)
 (c1 = "c", c2 = nothing)
 (c1 = "d", c2 = false)
 (c1 = "e", c2 = "D8")
 (c1 = nothing, c2 = "D8/0")
 (c1 = nothing, c2 = nothing)

YaXLSX.xl_columntable — Function

xl_columntable(sheet::Sheet, s::Union{UnitRange, AbstractString}; column_labels=nothing) -> ColumnTable
xl_columntable(sheet::Sheet; column_labels=nothing) -> ColumnTable

Get a ColumnTable from Sheet by cells or columns range.

xl_book = parse_xlsx(xlsx_two_sheets_table())

xl_sheet = xl_sheets(xl_book, 2)

julia> xl_columntable(xl_sheet, "1:2")
(
    A = Any["Numbers", 1.0, 2.0, 3.0, 4.0, 5.0, nothing, nothing], 
    B = Any["Names", "a", "b", "c", "d", "e", nothing, nothing]
)

julia> xl_columntable(xl_sheet, 3:4; column_labels=["C_col", "D_col"])
(
    C_col = Any[nothing, nothing, nothing, nothing, false, "D8", "D8/0", nothing], 
    D_col = Any[nothing, nothing, nothing, nothing, nothing, nothing, nothing, 100.0]
)